Added print_single_frame + better parsing of CLI CAN ID
